requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
openai==1.12.0
azure-search-documents==11.4.0
azure-cosmos==4.5.1
python-dotenv==1.0.0
pydantic==2.6.0